refactor: address code review findings for test detection and panic interpolation
This commit addresses several code quality improvements identified during code review:

1. Extract duplicate harness test detection logic to shared infrastructure
   - Move `collect_harness_test_functions` and `is_harness_test_descriptor` from
     both `no_unwrap_or_else_panic` and `no_expect_outside_tests` to `whitaker::hir`
   - Eliminates ~130 lines of duplicated code across the two lint crates
   - Provides a single source of truth for test harness detection logic

2. Fix early return behavior in PanicDetector visitor
   - Remove early return when first panic is found
   - Use OR-assignment to accumulate interpolation detection across multiple panics
   - Ensures consistent detection regardless of HIR traversal order
   - Changes `self.uses_interpolation = ...` to `self.uses_interpolation |= ...`

3. Consolidate policy tests using rstest parameterization
   - Replace 8 individual test functions with single parameterized test
   - Reduce test code from ~80 lines to ~30 lines
   - Maintain clear test case names via rstest case attributes
   - Improve maintainability and reduce duplication

4. Improve documentation accuracy
   - Update policy.rs doc comment to correctly describe interpolation detection
   - Replace incorrect reference to `_fmt`/`_display` entry points with accurate
     description of `fmt::Arguments::new_v1(_formatted)` construction detection

All validation checks pass:
- make check-fmt ✓
- make lint ✓
- make typecheck ✓
- make test ✓ (1244 tests passed, 2 skipped)
