+++ crates/no_unwrap_or_else_panic/src/driver.rs
@@ -9,7 +9,6 @@ use log::debug;
 use rustc_hir as hir;
 use rustc_hir::ExprKind;
 use rustc_lint::{LateContext, LateLintPass};
-use rustc_span::Symbol;
 use serde::Deserialize;
 use std::collections::HashSet;
 use whitaker::SharedConfig;
@@ -64,7 +63,7 @@ impl<'tcx> LateLintPass<'tcx> for NoUnwrapOrElsePanic {
 
         self.is_test_harness = cx.tcx.sess.opts.test;
         self.harness_test_functions = if self.is_test_harness {
-            collect_harness_test_functions(cx)
+            whitaker::hir::collect_harness_test_functions(cx)
         } else {
             HashSet::new()
         };
@@ -138,76 +137,6 @@ fn is_inside_harness_test_function<'tcx>(
             }
             _ => false,
         })
-}
-
-/// Collects all functions that the `rustc --test` harness identifies as tests.
-///
-/// The test harness synthesises a sibling `const` descriptor with the same name
-/// and source span as each test function. This function scans for those
-/// descriptors to recover test-function identity after the harness has consumed
-/// the original `#[test]` attributes.
-fn collect_harness_test_functions<'tcx>(cx: &LateContext<'tcx>) -> HashSet<hir::HirId> {
-    let root_items: Vec<_> = cx
-        .tcx
-        .hir_crate_items(())
-        .free_items()
-        .map(|id| cx.tcx.hir_item(id))
-        .collect();
-    let mut marked = HashSet::new();
-    collect_in_item_group(cx, &root_items, &mut marked);
-    marked
-}
-
-fn collect_in_item_group<'tcx>(
-    cx: &LateContext<'tcx>,
-    items: &[&'tcx hir::Item<'tcx>],
-    marked: &mut HashSet<hir::HirId>,
-) {
-    for item in items
-        .iter()
-        .copied()
-        .filter(|item| matches!(item.kind, hir::ItemKind::Fn { .. }))
-    {
-        let Some(ident) = item.kind.ident() else {
-            continue;
-        };
-
-        if items
-            .iter()
-            .any(|sibling| is_test_descriptor(item.hir_id(), ident.name, item.span, sibling))
-        {
-            marked.insert(item.hir_id());
-        }
-    }
-
-    // Recurse into submodules.
-    for item in items {
-        let hir::ItemKind::Mod(_, module) = item.kind else {
-            continue;
-        };
-        let module_items: Vec<_> = module
-            .item_ids
-            .iter()
-            .map(|id| cx.tcx.hir_item(*id))
-            .collect();
-        collect_in_item_group(cx, &module_items, marked);
-    }
-}
-
-/// The `--test` harness synthesises a `const` with the same name and source
-/// range as the test function.
-fn is_test_descriptor(
-    fn_hir_id: hir::HirId,
-    fn_name: Symbol,
-    fn_span: rustc_span::Span,
-    sibling: &hir::Item<'_>,
-) -> bool {
-    sibling.hir_id() != fn_hir_id
-        && matches!(sibling.kind, hir::ItemKind::Const(..))
-        && sibling
-            .kind
-            .ident()
-            .is_some_and(|ident| ident.name == fn_name && sibling.span.source_equal(fn_span))
 }
 
 fn load_configuration() -> Config {
