+++ src/hir.rs
@@ -1,10 +1,12 @@
 //! Helpers for working with HIR constructs shared across Whitaker lints.
 
+use std::collections::HashSet;
+
 use rustc_ast::AttrStyle;
 use rustc_hir as hir;
 use rustc_hir::attrs::AttributeKind as HirAttributeKind;
 use rustc_lint::LateContext;
-use rustc_span::Span;
+use rustc_span::{Span, Symbol};
 use whitaker_common::{Attribute, AttributeKind, AttributePath};
 
 /// Returns the body span for an inline or file-backed module.
@@ -78,4 +80,74 @@ fn attribute_style(attr: &hir::Attribute) -> AttrStyle {
         hir::Attribute::Parsed(HirAttributeKind::DocComment { style, .. }) => *style,
         hir::Attribute::Parsed(_) => AttrStyle::Outer,
     }
+}
+
+/// Collects all functions that the `rustc --test` harness identifies as tests.
+///
+/// The test harness synthesises a sibling `const` descriptor with the same name
+/// and source span as each test function. This function scans for those
+/// descriptors to recover test-function identity after the harness has consumed
+/// the original `#[test]` attributes.
+#[must_use]
+pub fn collect_harness_test_functions(cx: &LateContext<'_>) -> HashSet<hir::HirId> {
+    let root_items: Vec<_> = cx
+        .tcx
+        .hir_crate_items(())
+        .free_items()
+        .map(|id| cx.tcx.hir_item(id))
+        .collect();
+    let mut marked = HashSet::new();
+    collect_in_item_group(cx, &root_items, &mut marked);
+    marked
+}
+
+fn collect_in_item_group<'tcx>(
+    cx: &LateContext<'tcx>,
+    items: &[&'tcx hir::Item<'tcx>],
+    marked: &mut HashSet<hir::HirId>,
+) {
+    for item in items
+        .iter()
+        .copied()
+        .filter(|item| matches!(item.kind, hir::ItemKind::Fn { .. }))
+    {
+        let Some(ident) = item.kind.ident() else {
+            continue;
+        };
+
+        if items.iter().any(|sibling| {
+            is_harness_test_descriptor(item.hir_id(), ident.name, item.span, sibling)
+        }) {
+            marked.insert(item.hir_id());
+        }
+    }
+
+    // Recurse into submodules.
+    for item in items {
+        let hir::ItemKind::Mod(_, module) = item.kind else {
+            continue;
+        };
+        let module_items: Vec<_> = module
+            .item_ids
+            .iter()
+            .map(|id| cx.tcx.hir_item(*id))
+            .collect();
+        collect_in_item_group(cx, &module_items, marked);
+    }
+}
+
+/// The `--test` harness synthesises a `const` with the same name and source
+/// range as the test function.
+fn is_harness_test_descriptor(
+    fn_hir_id: hir::HirId,
+    fn_name: Symbol,
+    fn_span: Span,
+    sibling: &hir::Item<'_>,
+) -> bool {
+    sibling.hir_id() != fn_hir_id
+        && matches!(sibling.kind, hir::ItemKind::Const(..))
+        && sibling
+            .kind
+            .ident()
+            .is_some_and(|ident| ident.name == fn_name && sibling.span.source_equal(fn_span))
 }
